lesson_10/UnitTest/AccessRepo.java
package UnitTest;

public class AccessRepo {
    User user;
    Group group;

    public AccessRepo(User user, Group group) {
        this.user = user;
        this.group = group;
    }

    public User getUser() {
        return user;
    }

    public void setUser(User user) {
        this.user = user;
    }

    public Group getGroup() {
        return group;
    }

    public void setGroup(Group group) {
        this.group = group;
    }

    public User findUser(){
        return null;
    }
}
